更新灰度传感器中间件，优化里程计结构

+++ 03_Config/board_types.h
@@ -15,8 +15,8 @@ typedef signed short   int16;
 typedef unsigned short uint16;
 typedef signed int     int32;
 typedef unsigned int   uint32;
-typedef signed long    int64;   /* TODO: 选定芯片/编译器后确认 long 位宽，必要时改 long long */
-typedef unsigned long  uint64;
+typedef signed long long   int64;
+typedef unsigned long long uint64;
 
 /* 浮点 */
 typedef float          float32;
